refactor(Twitter): Major code refactor for piko settings
