+++ app/src/main/java/app/revanced/integrations/twitter/Utils.java
@@ -15,8 +15,8 @@ import app.revanced.integrations.shared.StringRef;
 import app.revanced.integrations.shared.settings.BooleanSetting;
 import app.revanced.integrations.shared.settings.StringSetting;
 import app.revanced.integrations.shared.settings.preference.SharedPrefCategory;
-import app.revanced.integrations.twitter.settings.BackupPrefFragment;
-import app.revanced.integrations.twitter.settings.RestorePrefFragment;
+import app.revanced.integrations.twitter.settings.fragments.BackupPrefFragment;
+import app.revanced.integrations.twitter.settings.fragments.RestorePrefFragment;
 import app.revanced.integrations.twitter.settings.Settings;
 import com.google.android.material.tabs.TabLayout$g;
 import org.json.JSONArray;
@@ -58,11 +58,11 @@ public class Utils {
         startActivityFromClassName(className);
     }
 
-    public static void startBackupActivity(boolean featureFlag) {
-        Intent intent = new Intent(ctx, BackupPrefFragment.class);
+    public static void startBackupActivity(Context context,boolean featureFlag) {
+        Intent intent = new Intent(context, BackupPrefFragment.class);
         intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         intent.putExtra("featureFlag", featureFlag);
-        ctx.startActivity(intent);
+        context.startActivity(intent);
     }
 
     public static void startRestoreActivity(boolean featureFlag) {
