+++ app/src/main/java/app/revanced/integrations/twitter/settings/ActivityHook.java
@@ -3,11 +3,13 @@ package app.revanced.integrations.twitter.settings;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-//import android.util.Log;
+import android.app.Fragment;
+import android.app.FragmentTransaction;
 import androidx.appcompat.widget.Toolbar;
 import app.revanced.integrations.shared.Utils;
 import android.app.Fragment;
 import app.revanced.integrations.twitter.settings.featureflags.FeatureFlagsFragment;
+import app.revanced.integrations.twitter.settings.fragments.*;
 
 import static app.revanced.integrations.shared.Utils.context;
 
@@ -15,33 +17,41 @@ import static app.revanced.integrations.shared.Utils.context;
 public class ActivityHook {
     public static Toolbar toolbar;
     private static final String EXTRA_PIKO_SETTINGS = "piko_settings";
-    private static final String EXTRA_FEATURE_FLAGS = "piko_feature_flags";
 
     public static boolean create(Activity act) {
         Fragment fragment = null;
-        if (act.getIntent().getBooleanExtra(EXTRA_PIKO_SETTINGS, false)) {
+        Intent intent = act.getIntent();
+        boolean addToBackStack = false;
+        if (intent.getBooleanExtra(EXTRA_PIKO_SETTINGS, false)) {
             fragment = new SettingsFragment();
-        }else if (act.getIntent().getBooleanExtra(EXTRA_FEATURE_FLAGS, false)) {
+        }else if (intent.getBooleanExtra(Settings.MISC_FEATURE_FLAGS.key, false)) {
             fragment = new FeatureFlagsFragment();
+        }else if (intent.getBooleanExtra(Settings.PATCH_INFO.key, false)) {
+            fragment = new SettingsAboutFragment();
         }
 
         if(fragment!=null) {
-           // Log.d("BRUH", "Activity requested");
-
-            act.setContentView(Utils.getResourceIdentifier("preference_fragment_activity", "layout"));
-            toolbar = act.findViewById(Utils.getResourceIdentifier("toolbar", "id"));
-            toolbar.setNavigationIcon(Utils.getResourceIdentifier("ic_vector_arrow_left", "drawable"));
-            toolbar.setTitle(Utils.getResourceString("piko_title_settings"));
-            toolbar.setNavigationOnClickListener(view -> act.onBackPressed());
-
-            act.getFragmentManager().beginTransaction().replace(Utils.getResourceIdentifier("fragment_container", "id"), fragment).commit();
-
+            startFragment(act,fragment,addToBackStack);
             return true;
         }
         return false;
     }
 
-    private static void startActivity(String activity_name)throws Exception{
+    public static void startFragment(Activity act, Fragment fragment,boolean addToBackStack){
+        act.setContentView(Utils.getResourceIdentifier("preference_fragment_activity", "layout"));
+        toolbar = act.findViewById(Utils.getResourceIdentifier("toolbar", "id"));
+        toolbar.setNavigationIcon(Utils.getResourceIdentifier("ic_vector_arrow_left", "drawable"));
+        toolbar.setTitle(Utils.getResourceString("piko_title_settings"));
+        toolbar.setNavigationOnClickListener(view -> act.onBackPressed());
+
+        FragmentTransaction transaction = act.getFragmentManager().beginTransaction().replace(Utils.getResourceIdentifier("fragment_container", "id"), fragment);
+        if(addToBackStack){
+            transaction.addToBackStack(null);
+        }
+        transaction.commit();
+    }
+
+    public static void startActivity(String activity_name)throws Exception{
         Intent intent = new Intent(context, Class.forName("com.twitter.android.AuthorizeAppActivity"));
         Bundle bundle = new Bundle();
         bundle.putBoolean(activity_name, true);
@@ -50,11 +60,17 @@ public class ActivityHook {
         context.startActivity(intent);
     }
 
-    public static void startSettingsActivity() throws Exception {
-        startActivity(EXTRA_PIKO_SETTINGS);
+    public static void startActivity(String activity_name,Bundle bundle)throws Exception{
+        Intent intent = new Intent(context, Class.forName("com.twitter.android.AuthorizeAppActivity"));
+        bundle.putBoolean(activity_name, true);
+        intent.putExtras(bundle);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        context.startActivity(intent);
     }
 
-    public static void startFeatureFlagsActivity() throws Exception {
-        startActivity(EXTRA_FEATURE_FLAGS);
+
+
+    public static void startSettingsActivity() throws Exception {
+        startActivity(EXTRA_PIKO_SETTINGS);
     }
 }
