+++ app/src/main/java/app/revanced/integrations/twitter/settings/widgets/EditTextPref.java
@@ -1,0 +1,38 @@
+package app.revanced.integrations.twitter.settings.widget;
+
+import android.content.Context;
+import android.preference.EditTextPreference;
+import android.util.AttributeSet;
+import app.revanced.integrations.twitter.Utils;
+import android.preference.Preference;
+
+public class EditTextPref extends EditTextPreference {
+    private static Helper helper;
+    public EditTextPref(Context context) {
+        super(context);
+        helper = new Helper(context);
+        init();
+    }
+    public EditTextPref(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        helper = new Helper(context);
+        init();
+    }
+
+    public EditTextPref(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        helper = new Helper(context);
+        init();
+    }
+
+    private void init() {
+        setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
+            @Override
+            public boolean onPreferenceChange(Preference preference, Object newValue) {
+                helper.setValue(preference,newValue);
+                preference.setSummary((String) newValue);
+                return true;
+            }
+        });
+    }
+}
